Adding file with 2 script to scan website
This script has 2 scripts in it.  One is basic with no error handling or loops or if conditions.  Second one has error handling and loops and if conditions.  

Both do the same thing: scan a target at port 80 and return the response by decoding it.   Lot of potential to expand.